Nuevos campos en el articulo (slug, image, draft, public y user), se mejoro el control de las publicaciones y las restricciones de los usuarios. La API de articulos se accede por medio del campo slug.

+++ articles/views.py
@@ -1,8 +1,12 @@
 # coding=utf-8
+from datetime import datetime
 from django.contrib import messages
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
+from django.db.models.query_utils import Q
 from django.http import HttpResponse, HttpResponseRedirect
+from django.http.response import Http404
 from django.shortcuts import render, get_object_or_404, redirect
+from django.utils import timezone
 
 # request.user.is_authenticated() - Determina si el usuario esta autenticado
 from articles.forms import ArticleForm
@@ -10,8 +14,22 @@ from articles.models import Article
 
 
 def article_list(request):
-    queryset_list = Article.objects.all()
-    paginator = Paginator(queryset_list, 6)
+    today = timezone.now()
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Article.objects.all()
+    else:
+        queryset_list = Article.objects.active()
+
+    query = request.GET.get("q")
+    if query:
+        queryset_list = queryset_list.filter(
+            Q(title__icontains=query) |
+            Q(content__icontains=query) |
+            Q(user__first_name__icontains=query) |
+            Q(user__last_name__icontains=query)
+        ).distinct()
+
+    paginator = Paginator(queryset_list, 10)
     page_request_var = 'page'
     page = request.GET.get(page_request_var)
     try:
@@ -24,13 +42,18 @@ def article_list(request):
     context = {
         "title": "List",
         "object_list": queryset,
-        "page_request_var": page_request_var
+        "page_request_var": page_request_var,
+        "today": today
     }
     return render(request, "article/list.html", context)
 
 
-def article_detail(request, id=None):
-    instance = get_object_or_404(Article, id=id)
+def article_detail(request, slug=None):
+    instance = get_object_or_404(Article, slug=slug)
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
+
     context = {
         "title": "Detail",
         "instance": instance
@@ -39,10 +62,15 @@ def article_detail(request, id=None):
 
 
 def article_create(request):
-    form = ArticleForm(request.POST or None)
+
+    if not request.user.is_authenticated():
+        raise Http404
+
+    form = ArticleForm(request.POST or None, request.FILES or None)
     if request.method == "POST":
         if form.is_valid():
             instance = form.save(commit=False)
+            instance.user = request.user
             instance.save()
             messages.success(request, "<a href='#'>El artículo</a> se creo correctamente.", extra_tags='html_safe')
             return HttpResponseRedirect(instance.get_absolute_url())
@@ -55,9 +83,13 @@ def article_create(request):
     return render(request, "article/form.html", context)
 
 
-def article_update(request, id=None):
-    instance = get_object_or_404(Article, id=id)
-    form = ArticleForm(request.POST or None, instance=instance)
+def article_update(request, slug=None):
+
+    if not request.user.is_authenticated():
+        raise Http404
+
+    instance = get_object_or_404(Article, slug=slug)
+    form = ArticleForm(request.POST or None, request.FILES or None, instance=instance)
     if request.method == "POST":
         if form.is_valid():
             instance = form.save(commit=False)
@@ -74,8 +106,12 @@ def article_update(request, id=None):
     return render(request, "article/form.html", context)
 
 
-def article_delete(request, id=None):
-    instance = get_object_or_404(Article, id=id)
+def article_delete(request, slug=None):
+
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+
+    instance = get_object_or_404(Article, slug=slug)
     instance.delete()
     messages.success(request, "Se ha borrado correctamente.")
     return redirect("article:list")
